+++ agent/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [1.0.0-rc.2](https://github.com/qvest-digital/mxl-k8s/compare/agent/v1.0.0-rc.1...agent/v1.0.0-rc.2) (2026-05-19)
+
+
+### Bug Fixes
+
+* **agent:** reader pod can't open mirrored flow on KIND demo ([#40](https://github.com/qvest-digital/mxl-k8s/issues/40)) ([27ee1f9](https://github.com/qvest-digital/mxl-k8s/commit/27ee1f9d300f6dd223f03f2f0bf2eb3953e4829f))
+* **shim,agent,gateway:** close intent path and quiet reconciler noise ([#41](https://github.com/qvest-digital/mxl-k8s/issues/41)) ([36d6d88](https://github.com/qvest-digital/mxl-k8s/commit/36d6d883aab66565d90b7832c04c0cfe3cf0d116))
+
 ## [1.0.0-rc.1](https://github.com/qvest-digital/mxl-k8s/compare/agent/v1.0.0-rc.0...agent/v1.0.0-rc.1) (2026-05-18)